Docs: Make VM data persistence (or lack of) clearer
Previously it was easy to not realise that data outside of
`/var/lib/docker` and `/var/lib/boot2docker` would be lost after reboots
of the VM. We now also link to the relevant FAQ section from the main
README, so it's easier to work out how to make permanent modifications.

Fixes #912.