Align ephemeral port range with default windows ephemeral port range.
When docker exposes ports in uses the ephemeral port range.
See: http://docs.docker.com/articles/networking/#binding-ports
The default linux, and thereby boot2docker, range is 32768 to 61000.
Windows however uses a range of 49152 to 65535. When an exposed port is
in the range of 32768 to 49151 applications on the host may be unable
to connect because the port is outside the windows ephemeral port range.
This change aligns the boot2docker range to the windows defaults.