Add proc_events netlink API support
This allows a process with CAP_NET_ADMIN to ask the kernel to notify
processes when ever a process or thread/task is created. This is needed
for some monitoring and security tools to run.

It was added to debian mainline in 2012 in this bug
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=588200

I had to switch connector from a module to built-in otherwise during
`make oldconfig` it would disable PROC_EVENTS again.