Less permissive file modes for /Users mount
When mounting C:\Users and the like, at least on Windows (I'm not sure if this is the same if the host OS is OSX in which case this might need adjusting) all files are given mode `777`.  This is a problem for a few reasons that I listed in sagemath/docker-images#19

Setting `600` for files seems to make sense, since for directories under `/Users`/`C:\Users` most files should be private anyways.  I'm split on whether files should be executable or not, but ultimately going with not to be on the safe side.