Enable ntpd daemon
I am using docker-machines as nodes in our Jenkins CI, I noticed that the machines time keeps drifting, After digging into your code, I noticed that the `rc.d/ntpd` sets the Do Not Demonize flag `-n`. After a week or so of monitoring, the time sync seems to be working.
Was there a reason for this flag? If not, is this a possible fix?