Fix fsync on dirs in Parallels Shared Folders
Previously prl_fs didn't provide fsync handler for directory operations
(but for files did). As a result respective syscalls failed. That
caused, for example, problems with running PostgreSQL in containers over
database on a shared volume.

This patch is really a kludge to workaround the issue. And should be
removed when Parallels Tools with fixed prl_fs is incorporated.

Parallels/docker-machine-parallels#71