Adjust conditional "haveged" startup according to bwh's suggestion for d-i
See https://bugs.debian.org/923675#72; essentially if we have hardware RNG of some type, skip software RNG, otherwise check whether we have enough entropy to maybe not bother.

The latter part of this check probably isn't strictly necessary (it should be reasonably harmless to start haveged in all cases of missing hardware RNG), but we're a little different from the debian-installer use case.  In d-i, they're only talking about starting `haveged` for the duration of the installation process, where the instance we start will potentially affect our users more directly (more akin to if they were discussing adding `haveged` to the default set of installed packages where hardware RNG does not exist, so we want to be a tad bit more conservative about where/when we apply this).